feat(payouts,reports): refactor to use shared CustomDataCell and catalog deps
