+++ packages/domains/payouts/package.json
@@ -10,8 +10,9 @@
         "./domain": "./domain/src/index.ts"
     },
     "dependencies": {
-        "@adyen/bento-vue3": "^1.94.0",
-        "@adyen/ui-assets-icons-16": "^7.7.2",
+        "@adyen/bento-vue3": "catalog:",
+        "@adyen/ui-assets-icons-16": "catalog:",
+        "@adyen/ui-assets-icons-40": "catalog:",
         "@integration-components/composables-vue": "workspace:*",
         "@integration-components/core": "workspace:*",
         "@integration-components/hooks-preact": "workspace:*",
