+++ packages/domains/payouts/vue/src/PayoutDetails/components/PayoutData.vue
@@ -5,6 +5,7 @@ import { BentoTypography, BentoCard, BentoTag, BentoLink, BentoButtonActions, Be
 import type { BentoColumn, BentoDatagridDataItem } from '@adyen/bento-vue3';
 import type { IPayoutDetails } from '@integration-components/types';
 import { DATE_FORMAT_PAYOUT_DETAILS } from '@integration-components/utils';
+import { formatAmountWithCurrencyCode } from '@integration-components/core/Localization/amount/amount-util';
 import useTimezoneAwareDateFormatting from '@integration-components/composables-vue/useTimezoneAwareDateFormatting';
 import { getPayoutAdjustmentType, getPayoutFundsCapturedType } from '@integration-components/payouts/domain';
 import {
@@ -128,9 +129,7 @@ function formatPayoutDate(dateStr: string): string {
     return dateFormat(dateStr, DATE_FORMAT_PAYOUT_DETAILS);
 }
 
-function formatAmount(amount: { value: number; currency: string }): string {
-    return `${i18n.amount(amount.value, amount.currency, { hideCurrency: true })} ${amount.currency}`;
-}
+const formatAmount = (amount: { value: number; currency: string }) => formatAmountWithCurrencyCode(amount.value, i18n.locale, amount.currency);
 
 const fundsCapturedColumns = computed<BentoColumn[]>(() => [
     {
@@ -205,7 +204,7 @@ const subtractionsRows = computed<BentoDatagridDataItem[]>(() =>
         <BentoCard>
             <div :class="titleClass">
                 <div :class="PD_TITLE_CONTAINER_CLASS">
-                    <BentoTypography variant="subtitle" stronger>
+                    <BentoTypography variant="title" stronger>
                         {{ i18n.get('payouts.details.tags.netPayout') }}
                     </BentoTypography>
                     <BentoTag v-if="payoutInner.isSumOfSameDayPayouts" variant="blue" :label="i18n.get('payouts.details.tags.sameDaySum')" />
